requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.30.0
Django==3.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.6.0
djangorestframework==3.13.1
idna==3.4
Pillow==9.2.0
python-decouple==3.4
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
urllib3==1.26.16
